moved About dialog
